+++ src/pages/brand/brand-list/index.tsx
@@ -1,8 +1,5 @@
-import {useQuery} from '@tanstack/react-query'
-import {fetchBrands} from './queryfn'
-import {useState, useEffect} from 'react'
-import useClassifyKorean from './hook/use-classify-korean'
-import {Box, Stack, Typography, styled} from '@mui/material'
+import {useState, useEffect, Suspense} from 'react'
+import {Box, Stack, Typography, keyframes, styled} from '@mui/material'
 import Banner from '@components/base/banner'
 import BranchBrandList from './branch-brand-list.'
 
@@ -25,9 +22,6 @@ const Kor = [
 ]
 
 const BrandList = () => {
-  const {data: brands} = useQuery(['brands'], () => fetchBrands())
-
-  const classifyBrands = useClassifyKorean(brands)
 
   const [korClass, setKorClass] = useState('')
 
@@ -57,7 +51,9 @@ const BrandList = () => {
             </KorButton>
           ))}
         </Stack>
-        <BranchBrandList korClass={korClass} classifyBrands={classifyBrands} brands={brands}/>
+        <Suspense fallback={<Loading />}>
+        <BranchBrandList korClass={korClass}/>
+        </Suspense>
       </Container>
     </Box>
   )
@@ -80,8 +76,27 @@ const Title = styled(Typography)(() => ({
 const KorButton = styled('button')<{active: boolean}>(({active, theme}) => ({
   width: '70px',
   height: '70px',
-  border: `2px solid ${active ? theme.palette.primary.main : 'black'}`,
+  border: `1px solid ${active ? theme.palette.primary.main : 'black'}`,
+  borderRadius: '8px',
   fontFamily: 'Pretendard',
   fontSize: '24px',
 }))
+
+const blinkAnimation = keyframes`
+    0%,
+    100%{
+        background-color: #eee
+    }
+    50%{
+        background-color: #fff
+    }
+`
+
+const Loading = styled(Box)(() => ({
+  width: '1200px',
+  height: '100vh',
+  animation: `${blinkAnimation} 0.8s infinite linear`,
+  borderRadius: '10px',
+}))
+
